再スタートの例外スロー解消・モデルのフレーム位置の更新処理追加・プレイヤーの待機状態のアニメーション遷移の修正

+++ Src/Manager/SoundManager.h
@@ -163,7 +163,7 @@ private:
 	std::map<SRC, Sound> sounds_;
 
 	// Žå‰¹—Ê
-	int volumeMaster_;
+	float volumeMaster_;
 
 
 	/// <summary>
